+++ docs/src/dev-docs/changelog.rst
@@ -27,7 +27,7 @@ Unreleased
 Fixed
 #####
 
-- ``NativePET`` can now evaluate on single-atom structures without crashing 
+- ``NativePET`` can now evaluate on single-atom structures without crashing
 
 Version 2025.5 - 2025-04-13
 ---------------------------